Refactor to move math-helper out&tests
