+++ Cargo.toml
@@ -1,3 +1,6 @@
+[profile.dev]
+split-debuginfo = "unpacked"
+
 [profile.release]
 overflow-checks = true
 panic           = 'unwind'
@@ -8,4 +11,4 @@ lto = true
 codegen-units = 1
 
 [workspace]
-members = ['node/*', 'pallets/*', 'runtime/*', 'integration-tests']
+members = ['node/*', 'pallets/*', 'runtime/*', 'integration-tests','support']
