+++ support/src/lib.rs
@@ -13,4 +13,4 @@
 // limitations under the License.
 
 #![cfg_attr(not(feature = "std"), no_std)]
-pub mod f64;
+pub mod math_helper;
